Plotting scripts are just becoming more interesting and more beautiful

+++ mininet_topos/Multinode_pox_codinghost.py
@@ -33,7 +33,7 @@ with open("%s/Development/VAST-0.4.6/bin/Mininet.ini" % home_dir, 'r') as config
     data = config.readlines()
     NODE_COUNT = int(data[data.index('#NODE_COUNT;    // Nodes started in simulation\n')+1])
     print("NODE_COUNT", NODE_COUNT)
-    BW = (int)(data[data.index('#BW;            // Bandwidth limit [Mbps], 0 if inifinte\n')+1])
+    BW = (float)(data[data.index('#BW;            // Bandwidth limit [Mbps], 0 if inifinte\n')+1])
     print ("BW", BW)
     DELAY = (int)(data[data.index('#DELAY;         // Delay in MS\n')+1])
     print ("DELAY", DELAY)
